+++ BattleMages/Components/Spells/FrostShield.cs
@@ -75,7 +75,7 @@ namespace BattleMages
 
                 if (enemy != null)
                 {
-                    enemy.TakeDamage(Stats.Damage);
+                    enemy.TakeDamage((int)Stats.Damage);
                     GameWorld.SoundManager.PlaySound("iceshardsbreaking", volume: 0.7f);
 
                     GameWorld.Scene.RemoveObject(GameObject);
